[#385] Refine segment tabs as a wrapper to be reused in another page
